[REF] Reorganize the exporters into submodules
To prep for multiple versions of dcm2bids I've divided up
the exporters and the bids classes will now be imported based on
user environment.
