+++ datman/exporters/bids_legacy.py
@@ -1,181 +1,34 @@
-"""Functions to export data into different file formats and organizations.
+"""Export to bids format when using dcmbids versions below '3'.
 
-To allow datman to export to a new format make a subclass of SessionExporter
-or SeriesExporter depending on whether the new format requires data from
-a complete scan session or a single series, respectively. The new subclass
-should implement all abstract methods, including 'export' which does the
-actual work of generating outputs.
+For dcm2bids versions 3 and higher (or dcm2bids versions accessed via
+container) 'dcm2bids', 'Dcm2bids' and 'Acquisition' are not accessible so these
+exporters cannot be used.
 
-Also, ensure that subclasses define the 'type' attribute to be a short
-unique key that can be referenced in config files (e.g. 'nii').
+When using versions below '3' though, this exporter has advantages over the
+newer one. Namely, its outputs_exist() method can better check the actual
+contents of the folder against what we expect to have been exported (reducing
+manual intervention). It can also force dcm2bids to properly export repeat
+sessions into the same folder, where newer versions will simply ignore them.
 """
-from abc import ABC, abstractmethod
 from collections import OrderedDict
-from datetime import datetime
 from glob import glob
 from json import JSONDecodeError
 import logging
 import os
 import re
 
-import pydicom as dicom
+from datman.scanid import make_filename
+from datman.utils import (splitext, get_extension, write_json, read_json,
+                          filter_niftis, read_blacklist, get_relative_source)
 
-import datman.config
-import datman.dashboard
-import datman.scan
-from datman.exceptions import (UndefinedSetting, DashboardException,
-                               ConfigException)
-from datman.scanid import (parse_bids_filename, ParseException,
-                           make_filename, KCNIIdentifier)
-from datman.utils import (run, make_temp_directory, get_extension,
-                          filter_niftis, find_tech_notes, read_blacklist,
-                          get_relative_source, read_json, write_json,
-                          parse_err_file)
+from dcm2bids import dcm2bids, Dcm2bids
+from dcm2bids.sidecar import Acquisition
 
-
-try:
-    from dcm2bids import dcm2bids, Dcm2bids
-    from dcm2bids.sidecar import Acquisition
-except ImportError:
-    DCM2BIDS_FOUND = False
-else:
-    DCM2BIDS_FOUND = True
+from .base import SessionExporter
 
 logger = logging.getLogger(__name__)
 
-
-def get_exporter(key, scope="series"):
-    """Find an exporter class for a given key identifier.
-
-    Args:
-        key (:obj:`str`): The 'type' identifier of a defined exporter (e.g.
-            'nii').
-        scope (:obj:`str`, optional): Whether to search for a series or session
-            exporter. Defaults to 'series'.
-
-    Returns:
-        :obj:`datman.exporters.Exporter`: The Exporter subclass for the type,
-            if one is defined, or else None.
-    """
-    if scope == "series":
-        exp_set = SERIES_EXPORTERS
-    else:
-        exp_set = SESSION_EXPORTERS
-
-    try:
-        exporter = exp_set[key]
-    except KeyError:
-        logger.error(
-            f"Unrecognized format {key} for {scope}, no exporters found.")
-        return None
-    return exporter
-
-
-class Exporter(ABC):
-    """An abstract base class for all Exporters.
-    """
-
-    # Subclasses must define this
-    type = None
-
-    @classmethod
-    def get_output_dir(cls, session):
-        """Retrieve the exporter's output dir without needing an instance.
-        """
-        return getattr(session, f"{cls.type}_path")
-
-    @abstractmethod
-    def outputs_exist(self):
-        """Whether outputs have already been generated for this Exporter.
-
-        Returns:
-            bool: True if all expected outputs exist, False otherwise.
-        """
-
-    @abstractmethod
-    def needs_raw_data(self):
-        """Whether raw data must be downloaded for the Exporter.
-
-        Returns:
-            bool: True if raw data must be given, False otherwise. Note that
-                False may be returned if outputs already exist.
-        """
-
-    @abstractmethod
-    def export(self, raw_data_dir, **kwargs):
-        """Exports raw data to the current Exporter's format.
-
-        Args:
-            raw_data_dir (:obj:`str`): The directory that contains the
-                downloaded raw data.
-        """
-
-    def make_output_dir(self):
-        """Creates the directory where the Exporter's outputs will be stored.
-
-        Returns:
-            bool: True if directory exists (or isn't needed), False otherwise.
-        """
-        try:
-            os.makedirs(self.output_dir)
-        except FileExistsError:
-            pass
-        except AttributeError:
-            logger.debug(f"output_dir not defined for {self}")
-        except PermissionError:
-            logger.error(f"Failed to make output dir {self.output_dir} - "
-                         "PermissionDenied.")
-            return False
-        return True
-
-
-class SessionExporter(Exporter):
-    """A base class for exporters that take an entire session as input.
-
-    Subclasses should override __init__ (without changing basic input args)
-    and call super().__init__(config, session, experiment, **kwargs).
-
-    The init function for SessionExporter largely exists to define expected
-    input arguments and set some universally needed attributes.
-    """
-
-    def __init__(self, config, session, experiment, dry_run=False, **kwargs):
-        self.experiment = experiment
-        self.config = config
-        self.session = session
-        self.dry_run = dry_run
-
-    def __repr__(self):
-        fq_name = str(self.__class__).replace("<class '", "").replace("'>", "")
-        name = fq_name.rsplit(".", maxsplit=1)[-1]
-        return f"<{name} - {self.experiment.name}>"
-
-
-class SeriesExporter(Exporter):
-    """A base class for exporters that take a single series as input.
-    """
-
-    # Subclasses should set this
-    ext = None
-
-    def __init__(self, output_dir, fname_root, echo_dict=None, dry_run=False,
-                 **kwargs):
-        self.output_dir = output_dir
-        self.fname_root = fname_root
-        self.echo_dict = echo_dict
-        self.dry_run = dry_run
-
-    def outputs_exist(self):
-        return os.path.exists(
-            os.path.join(self.output_dir, self.fname_root + self.ext))
-
-    def needs_raw_data(self):
-        return not self.outputs_exist()
-
-    def __repr__(self):
-        fq_name = str(self.__class__).replace("<class '", "").replace("'>", "")
-        name = fq_name.rsplit(".", maxsplit=1)[-1]
-        return f"<{name} - {self.fname_root}>"
+__all__ = ["BidsExporter", "NiiLinkExporter"]
 
 
 class BidsExporter(SessionExporter):
@@ -274,7 +127,8 @@ class BidsExporter(SessionExporter):
                 src_acqs = []
                 for acq in local_parser.acquisitions:
                     sidecar = acq.srcSidecar
-                    if str(sidecar.data['SeriesNumber']) in [scan.series, "10" + scan.series]:
+                    if str(sidecar.data['SeriesNumber']) in [
+                            scan.series, "10" + scan.series]:
                         src_acqs.append(acq)
 
                 for src_acq in src_acqs:
@@ -354,7 +208,7 @@ class BidsExporter(SessionExporter):
             os.makedirs(os.path.dirname(dest_path))
 
         for found in glob(source_path + "*"):
-            _, ext = datman.utils.splitext(found)
+            _, ext = splitext(found)
             os.rename(found, dest_path + ext)
 
     def get_xnat_parser(self):
@@ -477,10 +331,11 @@ class BidsExporter(SessionExporter):
         if self.outputs_exist():
             return
 
-        if not DCM2BIDS_FOUND:
-            logger.info(f"Unable to export to {self.output_dir}, "
-                        "Dcm2Bids not found.")
-            return
+        # Was this ever needed? The class should never have been made.
+        # if not DCM2BIDS_FOUND:
+        #     logger.info(f"Unable to export to {self.output_dir}, "
+        #                 "Dcm2Bids not found.")
+        #     return
 
         if self.dry_run:
             logger.info(f"Dry run: Skipping bids export to {self.output_dir}")
@@ -489,7 +344,6 @@ class BidsExporter(SessionExporter):
         # Store user settings in case they change during export
         orig_force = self.force_dcm2niix
         orig_refresh = self.refresh
-
 
         if int(self.repeat) > 1:
             # Must force dcm2niix export if it's a repeat.
@@ -610,7 +464,8 @@ class BidsExporter(SessionExporter):
         xnat_map = {}
         for acq in xnat_parser.acquisitions:
             try:
-                xnat_map.setdefault(acq.srcSidecar.scan, []).append(acq.dstRoot)
+                xnat_map.setdefault(acq.srcSidecar.scan, []).append(
+                    acq.dstRoot)
             except AttributeError:
                 # acqs belonging to previous sessions don't have
                 # srcSidecar.scan and should not be in xnat_map
@@ -1162,601 +1017,6 @@ class NiiLinkExporter(SessionExporter):
                 pass
             except Exception as exc:
                 logger.error(f"Failed to create {target}. Reason - {exc}")
-
-
-class DBExporter(SessionExporter):
-    """Add a datman-style session and its contents to datman's QC dashboard.
-    """
-
-    type = "db"
-
-    def __init__(self, config, session, experiment, **kwargs):
-        try:
-            study_resource_dir = config.get_path("resources")
-        except UndefinedSetting:
-            study_resource_dir = ""
-
-        try:
-            resources_dir = os.path.join(
-                config.get_path("resources"),
-                session._ident.get_full_subjectid_with_timepoint_session()
-            )
-        except UndefinedSetting:
-            resources_dir = ""
-
-        self.nii_path = session.nii_path
-        self.output_dir = None
-        self.ident = session._ident
-        self.study_resource_path = study_resource_dir
-        self.resources_path = resources_dir
-        self.date = experiment.date
-        super().__init__(config, session, experiment, **kwargs)
-
-    @property
-    def names(self):
-        """Gets list of valid datman-style scan names for a session.
-
-        Returns:
-            :obj:`dict`: A dictionary of datman style scan names mapped to
-                the bids style name if one can be found, otherwise, an
-                empty string.
-        """
-        names = {}
-        # use experiment.scans, so dashboard can report scans that didnt export
-        for scan in self.experiment.scans:
-            for name in scan.names:
-                names[name] = self.get_bids_name(name, self.session)
-
-        # Check the actual folder contents as well, in case symlinked scans
-        # exist that werent named on XNAT
-        for nii in self.session.niftis:
-            fname = nii.file_name.replace(nii.ext, "")
-            if fname in names:
-                continue
-            names[fname] = self.get_bids_name(fname, self.session)
-
-        return names
-
-    def get_bids_name(self, dm_name, session):
-        """Get BIDS style scan name from a datman style nifti.
-
-        Returns:
-            str: A valid bids style file name or an empty string if one
-                cannot be found.
-        """
-        found = [item for item in session.find_files(dm_name)
-                 if ".nii.gz" in item]
-        if not found or not os.path.islink(found[0]):
-            return ""
-        bids_src = os.readlink(found[0])
-        bids_name = os.path.basename(bids_src)
-        return bids_name.replace(get_extension(bids_name), "")
-
-    def export(self, *args, **kwargs):
-        if self.dry_run:
-            logger.info("Dry run: Skipping database update for "
-                        f"{str(self.ident)}")
-            return
-
-        if not datman.dashboard.dash_found:
-            logger.warning("Dashboard database not found, unable to add "
-                           f"{str(self.ident)} and its contents.")
-            return
-
-        session = self.make_session()
-
-        if not session.tech_notes and session.expects_notes():
-            self.add_tech_notes(session)
-
-        for file_stem in self.names:
-            self.make_scan(file_stem)
-
-    def outputs_exist(self):
-        try:
-            session = datman.dashboard.get_session(self.ident)
-        except DashboardException:
-            return False
-        except ParseException:
-            logger.error(
-                f"Session name {self.ident} is not datman format. Ignoring.")
-            return True
-
-        if not session:
-            return False
-
-        if not session.tech_notes and session.expects_notes():
-            return False
-
-        for name in self.names:
-            try:
-                scan = datman.dashboard.get_scan(name)
-            except DashboardException:
-                return False
-            except ParseException:
-                logger.error(
-                    f"Scan name {name} is not datman format. Ignoring.")
-                continue
-
-            if not scan:
-                return False
-
-            if self.errors_outdated(scan, name):
-                return False
-
-        return True
-
-    @classmethod
-    def get_output_dir(cls, session):
-        return None
-
-    def needs_raw_data(self):
-        return False
-
-    def make_session(self):
-        """Add the current session to datman's QC database.
-
-        Returns:
-            :obj:`dashboard.models.Session`: The created scan session or None.
-        """
-        logger.debug(f"Adding session {str(self.ident)} to dashboard.")
-        try:
-            session = datman.dashboard.get_session(self.ident, create=True)
-        except datman.dashboard.DashboardException as exc:
-            logger.error(f"Failed adding session {str(self.ident)} to "
-                         f"database. Reason: {exc}")
-            return None
-
-        self._set_alt_ids(session)
-        self._set_date(session)
-
-        return session
-
-    def _set_alt_ids(self, session):
-        """Add alternate ID formats for the scan session to the database.
-
-        Args:
-            session (:obj:`dashboard.models.Session`): A valid QC dashboard
-                scan session.
-        """
-        session.timepoint.bids_name = self.ident.get_bids_name()
-        session.timepoint.bids_session = self.ident.timepoint
-        session.save()
-
-        if not isinstance(self.ident, KCNIIdentifier):
-            return
-
-        session.timepoint.kcni_name = self.ident.get_xnat_subject_id()
-        session.kcni_name = self.ident.get_xnat_experiment_id()
-        session.save()
-        return
-
-    def _set_date(self, session):
-        """Add the scan date for a scan session to the QC database.
-
-        Args:
-            session (:obj:`dashboard.models.Session`): A valid QC dashboard
-                scan session.
-        """
-        if not self.date:
-            logger.debug(f"No scan date found for {str(self.ident)}, "
-                         "leaving blank.")
-            return
-
-        try:
-            date = datetime.strptime(self.date, '%Y-%m-%d')
-        except ValueError:
-            logger.error(f"Invalid scan date {self.date} for session "
-                         f"{str(self.ident)}")
-            return
-
-        if date == session.date:
-            return
-
-        session.date = date
-        session.save()
-
-    def add_tech_notes(self, session):
-        """Add the path to a scan session's tech notes to the database.
-
-        Args:
-            session (:obj:`dashboard.models.Session`): A valid QC dashboard
-                scan session.
-        """
-        notes = find_tech_notes(self.resources_path)
-        if not notes:
-            logger.debug(f"No tech notes found in {self.resources_path}")
-            return
-
-        # Store only the path relative to the resources dir
-        session.tech_notes = notes.replace(
-            self.study_resource_path, "").lstrip("/")
-        session.save()
-
-    def make_scan(self, file_stem):
-        """Add a single scan to datman's QC dashboard.
-
-        Args:
-            file_stem (:obj:`str`): A valid datman-style file name.
-        """
-        logger.debug(f"Adding scan {file_stem} to dashboard.")
-        try:
-            scan = datman.dashboard.get_scan(file_stem, create=True)
-        except datman.dashboard.DashboardException as exc:
-            logger.error(f"Failed adding scan {file_stem} to dashboard "
-                         f"with error: {exc}")
-            return
-        if self.experiment.is_shared():
-            source_session = self._get_source_session()
-            self._make_linked(scan, source_session)
-        self._add_bids_scan_name(scan, file_stem)
-        self._add_side_car(scan, file_stem)
-        self._update_conversion_errors(scan, file_stem)
-
-    def _make_linked(self, scan, source_session):
-        try:
-            source_session = datman.dashboard.get_session(source_session)
-        except datman.dashboard.DashboardException as exc:
-            logger.error(
-                f"Failed to link shared scan {scan} to source "
-                f"{source_session}. Reason - {exc}"
-            )
-            return
-        matches = [
-            source_scan for source_scan in source_session.scans
-            if (source_scan.series == scan.series and
-                source_scan.tag == scan.tag)
-        ]
-        if not matches or len(matches) > 1:
-            logger.error(
-                f"Failed to link shared scan {scan} to {source_session}."
-                " Reason - Unable to find source scan database record."
-            )
-            return
-
-        scan.source_id = matches[0].id
-        scan.save()
-
-    def _get_source_session(self):
-        """Get the ID of the source experiment for a shared XNATExperiment."""
-        try:
-            config = datman.config.config(study=self.experiment.source_name)
-        except ConfigException:
-            return self.experiment.source_name
-
-        try:
-            id_map = config.get_key('IdMap')
-        except UndefinedSetting:
-            return self.experiment.source_name
-
-        return str(datman.scanid.parse(self.experiment.source_name, id_map))
-
-    def _add_bids_scan_name(self, scan, dm_stem):
-        """Add a bids format file name to a series in the QC database.
-
-        Args:
-            scan (:obj:`dashboard.models.Scan`): A QC dashboard scan.
-            dm_stem (:obj:`str`): A valid bids format scan name, or an
-                empty string if the update should be skipped.
-        """
-        bids_stem = self.names[dm_stem]
-        if not bids_stem:
-            return
-
-        try:
-            bids_ident = parse_bids_filename(bids_stem)
-        except ParseException:
-            logger.debug(f"Failed to parse bids file name {bids_stem}")
-            return
-        scan.add_bids(str(bids_ident))
-
-    def _add_side_car(self, scan, file_stem):
-        """Add the JSON side car contents to the QC database.
-
-        Args:
-            scan (:obj:`dashboard.models.Scan`): A QC dashboard scan.
-            file_stem (:obj:`str`): A valid datman-style file name. Used to
-                find the json side car file.
-        """
-        nii_file = self._get_file(file_stem, ".nii.gz")
-        if not nii_file:
-            # File exists on xnat but hasnt been generated.
-            return
-
-        side_car = self._get_file(file_stem, ".json")
-        if not side_car:
-            logger.error(f"Missing json side car for {file_stem}")
-            return
-
-        try:
-            scan.add_json(side_car)
-        except Exception as exc:
-            logger.error("Failed to add JSON side car to dashboard "
-                         f"record for {side_car}. Reason - {exc}")
-
-    def _update_conversion_errors(self, scan, file_stem):
-        """Add any dcm2niix conversion errors to the QC database.
-
-        Args:
-            scan (:obj:`dashboard.models.Scan`): A QC dashboard scan.
-            file_stem (:obj:`str`): A valid datman style file name. Used to
-                find the conversion error file (if one exists).
-        """
-        convert_errors = self._get_file(file_stem, ".err")
-        if not convert_errors:
-            if scan.conv_errors:
-                # Erase the error message from the DB, because it
-                # has been resolved.
-                scan.add_error(None)
-            return
-        message = self._read_file(convert_errors)
-        scan.add_error(message)
-
-    def _get_file(self, fname, ext):
-        """Find a file on the file system.
-
-        Args:
-            fname (:obj:`str`): A file name (minus extension).
-            ext (:obj:`str`): A file extension.
-
-        Returns:
-            str: The full path to the file matching the given name and
-                extension, otherwise None.
-        """
-        found = os.path.join(self.nii_path, fname + ext)
-        if not os.path.exists(found):
-            bl_found = os.path.join(self.nii_path, 'blacklisted', fname + ext)
-            if os.path.exists(bl_found):
-                return bl_found
-            logger.debug(f"File not found {found}")
-            return None
-        return found
-
-    def _read_file(self, fpath):
-        """Read the contents of a file.
-
-        Args:
-            fpath (:obj:`str`): The full path to a file.
-
-        Returns:
-            str: The contents of the file or None if the file cannot be read.
-        """
-        try:
-            with open(fpath, "r") as file_handle:
-                message = file_handle.readlines()
-        except Exception as exc:
-            logger.debug(f"Can't read file {fpath} - {exc}")
-            return None
-        return message
-
-    def errors_outdated(self, scan, fname):
-        err_file = self._get_file(fname, ".err")
-        if not err_file and scan.conv_errors:
-            # Error is resolved, but still appears in database
-            return True
-        if err_file and not scan.conv_errors:
-            # Error has appeared, but isnt recorded in database
-            return True
-        if err_file and scan.conv_errors:
-            # Error exists in both locations, but may have changed
-            message = self._read_file(err_file)
-            if isinstance(message, list):
-                message = "\n".join(message)
-            return message != scan.conv_errors
-        return False
-
-
-class NiiExporter(SeriesExporter):
-    """Export a series to nifti format with datman-style names.
-    """
-
-    ext = ".nii.gz"
-
-    type = "nii"
-
-    def export(self, raw_data_dir, **kwargs):
-        if self.dry_run:
-            logger.info(f"Dry run: Skipping export of {self.fname_root}")
-            return
-
-        if self.outputs_exist():
-            logger.debug(f"Outputs exist for {self.fname_root}, skipping.")
-            return
-
-        self.make_output_dir()
-
-        with make_temp_directory(prefix="export_nifti_") as tmp:
-            _, log_msgs = run(f'dcm2niix -z y -b y -o {tmp} {raw_data_dir}',
-                              self.dry_run)
-            for tmp_file in glob(f"{tmp}/*"):
-                self.move_file(tmp_file)
-                stem = self._get_fname(tmp_file)
-                self.report_issues(stem, str(log_msgs))
-
-    def move_file(self, gen_file):
-        """Move the temp outputs of dcm2niix to the intended output directory.
-
-        Args:
-            gen_file (:obj:`str`): The full path to the generated nifti file
-                to move.
-        """
-        fname = self._get_fname(gen_file)
-
-        if not fname:
-            return
-
-        out_file = os.path.join(self.output_dir, fname)
-        if os.path.exists(out_file):
-            logger.info(f"Output {out_file} already exists. Skipping.")
-            return
-
-        return_code, _ = run(f"mv {gen_file} {out_file}", self.dry_run)
-        if return_code:
-            logger.debug(f"Moving dcm2niix output {gen_file} to {out_file} "
-                         "has failed.")
-
-    def _get_fname(self, gen_file):
-        """Get the intended datman-style name for a generated file.
-
-        Args:
-            gen_file (:obj:`str`): The full path to the generated nifti file
-                to move.
-
-        Result:
-            str: A string filename (with extension) or an empty string.
-        """
-        ext = get_extension(gen_file)
-        bname = os.path.basename(gen_file)
-
-        if self.echo_dict:
-            stem = self._get_echo_fname(bname, ext)
-            if stem != self.fname_root:
-                # File belongs to the wrong echo, skip it
-                return ""
-        else:
-            stem = self.fname_root
-        return stem + ext
-
-    def _get_echo_fname(self, fname, ext):
-        """Get a valid datman-style file name from a multiecho file.
-
-        Args:
-            fname (:obj:`str`): A filename to parse for an echo number.
-            ext (:obj:`str`): The file extension to use.
-
-        Returns:
-            str: A valid datman-style file name or an empty string if one
-                cannot be made.
-        """
-        # Match a 14 digit timestamp and 1-3 digit series num
-        regex = "files_(.*)_([0-9]{14})_([0-9]{1,3})(.*)?" + ext
-        match = re.search(regex, fname)
-
-        if not match:
-            logger.error(f"Can't parse valid echo number from {fname}.")
-            return ""
-
-        try:
-            echo = int(match.group(4).split('e')[-1][0])
-            stem = self.echo_dict[echo]
-        except Exception:
-            logger.error(f"Can't parse valid echo number from {fname}")
-            return ""
-
-        return stem
-
-    def report_issues(self, stem, messages):
-        """Write an error log if dcm2niix had errors during conversion.
-
-        Args:
-            stem (:obj:`stem`): A valid datman-style file name (minus
-                extension).
-            messages (:obj:`str`): Error messages to write.
-        """
-        if self.dry_run:
-            logger.info(f"DRYRUN - Skipping write of error log for {stem}")
-            return
-
-        if 'missing images' not in messages:
-            # The only issue we care about currently is if files are missing
-            return
-
-        dest = os.path.join(self.output_dir, stem) + ".err"
-        self._write_error_log(dest, messages)
-
-    def _write_error_log(self, dest, messages):
-        """Write an error message to the file system.
-
-        Args:
-            dest (:obj:`str`): The full path of the file to write.
-            messages (:obj:`str`): Intended contents of the error log.
-        """
-        try:
-            with open(dest, "w") as output:
-                output.write(messages)
-        except Exception as exc:
-            logger.error(f"Failed writing dcm2niix errors to {dest}. "
-                         f"Reason - {type(exc).__name__} {exc} ")
-
-
-class DcmExporter(SeriesExporter):
-    """Export a single dicom from a scan.
-    """
-
-    type = "dcm"
-    ext = ".dcm"
-
-    def export(self, raw_data_dir, **kwargs):
-        self.make_output_dir()
-
-        if self.echo_dict:
-            self._export_multi_echo(raw_data_dir)
-            return
-
-        dcm_file = self._find_dcm(raw_data_dir)
-        if not dcm_file:
-            logger.error(f"No dicom files found in {raw_data_dir}")
-            return
-
-        logger.debug(f"Exporting a dcm file from {raw_data_dir} to "
-                     f"{self.output_dir}")
-        output = os.path.join(self.output_dir, self.fname_root + self.ext)
-        run(f"cp {dcm_file} {output}", self.dry_run)
-
-    def _find_dcm(self, raw_data_dir):
-        """Find the path to a valid dicom in the given directory.
-
-        Args:
-            raw_data_dir (:obj:`str`): The full path to the directory where
-                raw dicoms were downloaded for the series.
-
-        Returns:
-            str: the full path to the first readable dicom found.
-        """
-        for path in glob(f"{raw_data_dir}/*"):
-            try:
-                dicom.read_file(path)
-            except dicom.filereader.InvalidDicomError:
-                pass
-            else:
-                return path
-        return ""
-
-    def _export_multi_echo(self, raw_data_dir):
-        """Find a single valid dicom for each echo in a multiecho scan.
-
-        Args:
-            raw_data_dir (:obj:`str`): The full path to the directory where
-                raw dicoms were downloaded for the series.
-        """
-        dcm_dict = {}
-        for path in glob(f"{raw_data_dir}/*"):
-            try:
-                dcm_file = dicom.read_file(path)
-            except dicom.filereader.InvalidDicomError:
-                continue
-            dcm_echo_num = dcm_file.EchoNumbers
-            if dcm_echo_num not in dcm_dict:
-                dcm_dict[int(dcm_echo_num)] = path
-            if len(dcm_dict) == len(self.echo_dict):
-                break
-
-        for echo_num, dcm_echo_num in zip(self.echo_dict.keys(),
-                                          dcm_dict.keys()):
-            output_file = os.path.join(self.output_dir,
-                                       self.echo_dict[echo_num] + self.ext)
-            logger.debug(f"Exporting a dcm file from {raw_data_dir} to "
-                         f"{output_file}")
-            cmd = f"cp {dcm_dict[dcm_echo_num]} {output_file}"
-            run(cmd, self.dry_run)
-
-
-SESSION_EXPORTERS = {
-    exp.type: exp for exp in SessionExporter.__subclasses__()
-}
-
-SERIES_EXPORTERS = {
-    exp.type: exp for exp in SeriesExporter.__subclasses__()
-}
 
 
 class FakeSidecar(dcm2bids.Sidecar):
